refactor(subscriberinfo): move transform -> triggered
